bug fixed

+++ CPU_Simulator2/src/SimulationStats.java
@@ -5,37 +5,37 @@
  */
 public class SimulationStats {
 //All Overall Stats
-float TotalCreadtedProcesses;
-float CPUutilization;					//Total time cpu is used
-float switvhTime;
+double TotalCreadtedProcesses;
+double CPUutilization;					//Total time cpu is used
+double switvhTime;
 
 //All IO bound Stats
-float Total_I_O;
-float IO_Completed;
-float total_IO_Cpu_Time;
-float avg_IO_Cpu_Time;
-float total_IO_Ready_Wait;
-float avg_IO_Ready_Wait;
-float total_IO_Serve;
-float avg_IO_Serve;
-float total_IO_TurnAround;
-float avg_IO_TurnAround;
-float total_IO_Interrupts;
-float avg_IO_Interrupts;
+double Total_I_O;
+double IO_Completed;
+double total_IO_Cpu_Time;
+double avg_IO_Cpu_Time;
+double total_IO_Ready_Wait;
+double avg_IO_Ready_Wait;
+double total_IO_Serve;
+double avg_IO_Serve;
+double total_IO_TurnAround;
+double avg_IO_TurnAround;
+double total_IO_Interrupts;
+double avg_IO_Interrupts;
 
 //All CPU bound process Stats
-float Total_CPU;
-float CPU_Completed;
-float total_CPU_Cpu_Time;
-float avg_CPU_Cpu_Time;
-float total_CPU_Ready_Wait;
-float avg_CPU_Ready_Wait;
-float total_CPU_Serve;
-float avg_CPU_Serve;
-float total_CPU_TurnAround;
-float avg_CPU_TurnAround;
-float total_CPU_Interrupts;
-float avg_CPU_Interrupts;
+double Total_CPU;
+double CPU_Completed;
+double total_CPU_Cpu_Time;
+double avg_CPU_Cpu_Time;
+double total_CPU_Ready_Wait;
+double avg_CPU_Ready_Wait;
+double total_CPU_Serve;
+double avg_CPU_Serve;
+double total_CPU_TurnAround;
+double avg_CPU_TurnAround;
+double total_CPU_Interrupts;
+double avg_CPU_Interrupts;
 
 public SimulationStats() {
 	TotalCreadtedProcesses = 0;
@@ -58,53 +58,53 @@ public SimulationStats() {
 	avg_CPU_Interrupts = 0;
 }
 
-public float getAvg_IO_Serve() {
+public double getAvg_IO_Serve() {
 	return avg_IO_Serve;
 }
 
-public void setAvg_IO_Serve(float avg_IO_Serve) {
+public void setAvg_IO_Serve(double avg_IO_Serve) {
 	this.avg_IO_Serve = avg_IO_Serve;
 }
 
 /**
  * @param totalCreadtedProcesses the totalCreadtedProcesses to set
  */
-public void updateTotalCreadtedProcesses(float totalCreadtedProcesses) {
+public void updateTotalCreadtedProcesses(double totalCreadtedProcesses) {
 	TotalCreadtedProcesses = TotalCreadtedProcesses + totalCreadtedProcesses;
 }
 
 /**
  * @param cPUutilization the cPUutilization to set
  */
-public void updateCPUutilization(float cPUutilization) {
+public void updateCPUutilization(double cPUutilization) {
 	CPUutilization = CPUutilization + cPUutilization;
 }
 
 /**
  * @param switvhTime the switvhTime to set
  */
-public void updateSwitvhTime(float switvhTime) {
+public void updateSwitvhTime(double switvhTime) {
 	this.switvhTime = this.switvhTime + switvhTime;
 }
 
 /**
  * @param total_I_O the total_I_O to set
  */
-public void updateTotal_I_O(float total_I_O) {
+public void updateTotal_I_O(double total_I_O) {
 	Total_I_O = Total_I_O+ total_I_O;
 }
 
 /**
  * @param iO_Completed the iO_Completed to set
  */
-public void updateIO_Completed(float iO_Completed) {
+public void updateIO_Completed(double iO_Completed) {
 	IO_Completed = IO_Completed + iO_Completed;
 }
 
 /**
  * @param avg_IO_Cpu_Time the avg_IO_Cpu_Time to set
  */
-public void updateAvg_IO_Cpu_Time(float new_IO_Cpu_Time) {
+public void updateAvg_IO_Cpu_Time(double new_IO_Cpu_Time) {
 	this.total_IO_Cpu_Time = total_IO_Cpu_Time + new_IO_Cpu_Time;
 	this.avg_IO_Cpu_Time = this.total_IO_Cpu_Time/IO_Completed;
 }
@@ -112,28 +112,28 @@ public void updateAvg_IO_Cpu_Time(float new_IO_Cpu_Time) {
 /**
  * @param avg_IO_Ready_Wait the avg_IO_Ready_Wait to set
  */
-public void updateAvg_IO_Ready_Wait(float new_IO_Ready_Wait) {
+public void updateAvg_IO_Ready_Wait(double new_IO_Ready_Wait) {
 	this.total_IO_Ready_Wait = this.total_IO_Ready_Wait + new_IO_Ready_Wait;
 }
 
 /**
  * @param avg_IO_Serve the avg_IO_Serve to set
  */
-public void updateAvg_IO_Serve(float new_IO_Serve) {
+public void updateAvg_IO_Serve(double new_IO_Serve) {
 	this.total_IO_Serve = this.total_IO_Serve + new_IO_Serve;
 }
 
 /**
  * @param avg_IO_TurnAround the avg_IO_TurnAround to set
  */
-public void updateAvg_IO_TurnAround(float new_IO_TurnAround) {
+public void updateAvg_IO_TurnAround(double new_IO_TurnAround) {
 	this.total_IO_TurnAround = this.total_CPU_TurnAround + new_IO_TurnAround;
 }
 
 /**
  * @param avg_IO_Interrupts the avg_IO_Interrupts to set
  */
-public void updateAvg_IO_Interrupts(float new_IO_Interrupts) {
+public void updateAvg_IO_Interrupts(double new_IO_Interrupts) {
 	this.total_IO_Interrupts = this.total_IO_Interrupts + new_IO_Interrupts;
 	this.avg_IO_Interrupts = this.total_IO_Interrupts/IO_Completed;
 }
@@ -141,42 +141,42 @@ public void updateAvg_IO_Interrupts(float new_IO_Interrupts) {
 /**
  * @param total_CPU the total_CPU to set
  */
-public void updateTotal_CPU(float total_CPU) {
+public void updateTotal_CPU(double total_CPU) {
 	Total_CPU = Total_CPU+ total_CPU;
 }
 
 /**
  * @param cPU_Completed the cPU_Completed to set
  */
-public void updateCPU_Completed(float cPU_Completed) {
+public void updateCPU_Completed(double cPU_Completed) {
 	CPU_Completed = CPU_Completed + cPU_Completed;
 }
 
 /**
  * @param avg_CPU_Cpu_Time the avg_CPU_Cpu_Time to set
  */
-public void updateAvg_CPU_Cpu_Time(float new_CPU_Cpu_Time) {
+public void updateAvg_CPU_Cpu_Time(double new_CPU_Cpu_Time) {
 	this.total_CPU_Cpu_Time = this.total_CPU_Cpu_Time + new_CPU_Cpu_Time;
 }
 
 /**
  * @param avg_CPU_Ready_Wait the avg_CPU_Ready_Wait to set
  */
-public void updateAvg_CPU_Ready_Wait(float new_CPU_Ready_Wait) {
+public void updateAvg_CPU_Ready_Wait(double new_CPU_Ready_Wait) {
 	this.total_CPU_Ready_Wait = this.total_CPU_Ready_Wait + new_CPU_Ready_Wait;
 }
 
 /**
  * @param avg_CPU_Serve the avg_CPU_Serve to set
  */
-public void updateAvg_CPU_Serve(float new_CPU_Serve) {
+public void updateAvg_CPU_Serve(double new_CPU_Serve) {
 	this.total_CPU_Serve = this.total_CPU_Serve + new_CPU_Serve;
 }
 
 /**
  * @param avg_CPU_TurnAround the avg_CPU_TurnAround to set
  */
-public void setAvg_CPU_TurnAround(float new_CPU_TurnAround) {
+public void setAvg_CPU_TurnAround(double new_CPU_TurnAround) {
 	this.total_CPU_TurnAround = this.total_CPU_TurnAround + new_CPU_TurnAround;
 	this.avg_CPU_TurnAround = this.total_CPU_TurnAround/CPU_Completed;
 }
@@ -184,7 +184,7 @@ public void setAvg_CPU_TurnAround(float new_CPU_TurnAround) {
 /**
  * @param avg_CPU_Interrupts the avg_CPU_Interrupts to set
  */
-public void updateAvg_CPU_Interrupts(float new_CPU_Interrupts) {
+public void updateAvg_CPU_Interrupts(double new_CPU_Interrupts) {
 	this.total_CPU_Interrupts = this.total_CPU_Interrupts + new_CPU_Interrupts; 
 	this.avg_CPU_Interrupts = this.total_CPU_Interrupts/CPU_Completed;
 }
